Files moved into place under lib
