--- lib/features/support/values.rb
@@ -1,11 +1,0 @@
-# Provides a hash to store values for cross-request comparisons
-class Store
-  class << self
-    # Returns the hash of stored values for the current scenario
-    #
-    # @return [Hash] The stored value hash
-    def values
-      @values ||= {}
-    end
-  end
-end
