+++ lib/maze/browser_stack_utils.rb
