+++ lib/maze/capabilities/browsers.yml
