+++ lib/maze/capabilities/capabilities.rb
