+++ lib/maze/capabilities/devices.rb
