+++ lib/maze/compare.rb
