+++ lib/maze/configuration.rb
