+++ lib/maze/docker.rb
