+++ lib/maze/drivers/appium_driver.rb
