+++ lib/maze/drivers/resilient_driver.rb
