+++ lib/maze/drivers/selenium_driver.rb
