+++ lib/maze/hooks/hooks.rb
