+++ lib/maze/interactive_cli.rb
