+++ lib/maze/logger.rb
