+++ lib/maze/network.rb
