+++ lib/maze/options/option.rb
