+++ lib/maze/options/option_parser.rb
