+++ lib/maze/options/option_processor.rb
