+++ lib/maze/options/option_validator.rb
