+++ lib/maze/proxy.rb
