+++ lib/maze/runner.rb
